bot/requirements.txt
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0